+++ zanshinsdk/client.py
@@ -1,16 +1,12 @@
 from __future__ import annotations
 
 import logging
-import sys
-import time
 from configparser import RawConfigParser
 from enum import Enum
-from importlib.util import find_spec, module_from_spec
-from math import ceil
 from os import environ
 from os.path import isfile
 from pathlib import Path
-from typing import Dict, Iterable, Iterator, Optional, Union
+from typing import Optional, Union
 from urllib.parse import urlparse
 from uuid import UUID
 
@@ -437,1821 +433,6 @@ class Client:
             )
         response.raise_for_status()
         return response
-
-    ###################################################
-    # Organization Follower
-    ###################################################
-
-    def iter_organization_followers(
-        self, organization_id: Union[UUID, str]
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the followers of an organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationFollowers>
-        :param organization_id: the ID of the organization
-        :return: an iterator over the organization followers objects
-        """
-        yield from self._request(
-            "GET", f"/organizations/{validate_uuid(organization_id)}/followers"
-        ).json()
-
-    def stop_organization_follower(
-        self, organization_id: Union[UUID, str], follower_id: Union[UUID, str]
-    ) -> bool:
-        """
-        Stops one organization follower of another.
-        <https://api.zanshin.tenchisecurity.com/#operation/removeOrganizationFollower>
-        :param organization_id: the ID of the organization
-        :param follower_id: the ID of the follower
-        :return: a boolean if success
-        """
-        return self._request(
-            "DELETE",
-            f"/organizations/{validate_uuid(organization_id)}/followers/"
-            f"{validate_uuid(follower_id)}",
-        ).json()
-
-    ###################################################
-    # Organization Follower Request
-    ###################################################
-
-    def iter_organization_follower_requests(
-        self, organization_id: Union[UUID, str]
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the follower requests of an organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationFollowRequests>
-        :param organization_id: the ID of the organization
-        :return: an iterator over the organization follower requests objects
-        """
-        yield from self._request(
-            "GET", f"/organizations/{validate_uuid(organization_id)}/followers/requests"
-        ).json()
-
-    def create_organization_follower_request(
-        self, organization_id: Union[UUID, str], token: Union[UUID, str]
-    ) -> Dict:
-        """
-        Create organization follower request.
-        <https://api.zanshin.tenchisecurity.com/#operation/createOrganizationFollowRequests>
-        :param organization_id: the ID of the organization
-        :param token: the token of the follower request
-        :return: a dict representing the organization follower
-        """
-        body = {
-            "token": validate_uuid(token),
-        }
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/followers/requests",
-            body=body,
-        ).json()
-
-    def get_organization_follower_request(
-        self, organization_id: Union[UUID, str], token: Union[UUID, str]
-    ) -> Dict:
-        """
-        Get organization follower request.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationFollowRequestsByToken>
-        :param organization_id: the ID of the organization
-        :param token: the token of the follower request
-        :return: a dict representing the organization follower
-        """
-        return self._request(
-            "GET",
-            f"/organizations/{validate_uuid(organization_id)}/followers/requests/"
-            f"{validate_uuid(token)}",
-        ).json()
-
-    def delete_organization_follower_request(
-        self, organization_id: Union[UUID, str], follower_id: Union[UUID, str]
-    ) -> bool:
-        """
-        Delete organization follower request.
-        <https://api.zanshin.tenchisecurity.com/#operation/deleteOrganizationFollowRequestsbyToken>
-        :param organization_id: the ID of the organization
-        :param follower_id: the ID of the follower
-        :return: a boolean if success
-        """
-        return self._request(
-            "DELETE",
-            f"/organizations/{validate_uuid(organization_id)}/followers/requests/"
-            f"{validate_uuid(follower_id)}",
-        ).json()
-
-    ###################################################
-    # Organization Following
-    ###################################################
-
-    def iter_organization_following(
-        self, organization_id: Union[UUID, str]
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the following of an organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationFollowing>
-        :param organization_id: the ID of the organization whose followed organizations we should list
-        :return: an iterator over the JSON decoded followed organizations
-        """
-        yield from self._request(
-            "GET", f"/organizations/{validate_uuid(organization_id)}/following"
-        ).json()
-
-    def stop_organization_following(
-        self, organization_id: Union[UUID, str], following_id: Union[UUID, str]
-    ) -> bool:
-        """
-        Stops one organization following of another.
-        <https://api.zanshin.tenchisecurity.com/#operation/removeOrganizationFollowingById>
-        :param organization_id: the follower organization ID
-        :param following_id:  the followed organization ID
-        :return: a boolean indicating whether the operation was successful
-        """
-        return self._request(
-            "DELETE",
-            f"/organizations/{validate_uuid(organization_id)}/following/"
-            f"{validate_uuid(following_id)}",
-        ).json()
-
-    ###################################################
-    # Organization Following Request
-    ###################################################
-
-    def iter_organization_following_requests(
-        self, organization_id: Union[UUID, str]
-    ) -> Iterator[Dict]:
-        """
-        Returns all requests received by an organization to follow another.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationFollowingRequests>
-        :param organization_id: the ID of the organization that was invited to follow another
-        :return: an iterator over the JSON decoded following requests
-        """
-        yield from self._request(
-            "GET", f"/organizations/{validate_uuid(organization_id)}/following/requests"
-        ).json()
-
-    def get_organization_following_request(
-        self, organization_id: Union[UUID, str], following_id: Union[UUID, str]
-    ) -> Dict:
-        """
-        Returns a request received by an organization to follow another.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationFollowingRequestByToken>
-        :param organization_id: the ID of the organization
-        :param following_id: the ID of the following
-        :return: a dict representing the following request
-        """
-        return self._request(
-            "GET",
-            f"/organizations/{validate_uuid(organization_id)}/following/requests/"
-            f"{validate_uuid(following_id)}",
-        ).json()
-
-    def accept_organization_following_request(
-        self, organization_id: Union[UUID, str], following_id: Union[UUID, str]
-    ) -> Dict:
-        """
-        Accepts a request to follow another organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/acceptOrganizationFollowingRequestByToken>
-        :param organization_id: the ID of the organization who was invited to follow another
-        :param following_id: the ID of the organization who is going to be followed
-        :return: a decoded JSON object describing the newly established following relationship
-        """
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/following/requests/"
-            f"{validate_uuid(following_id)}/accept",
-        ).json()
-
-    def decline_organization_following_request(
-        self, organization_id: Union[UUID, str], following_id: Union[UUID, str]
-    ) -> Dict:
-        """
-        Declines a request to follow another organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/declineOrganizationFollowingRequestByToken>
-        :param organization_id: the ID of the organization who was invited to follow another
-        :param following_id: the ID of the organization who was going to be followed
-        :return: a decoded JSON object describing the newly established following relationship
-        """
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/following/requests/"
-            f"{validate_uuid(following_id)}/decline",
-        ).json()
-
-    ###################################################
-    # Organization Scan Target
-    ###################################################
-
-    def iter_organization_scan_targets(
-        self, organization_id: Union[UUID, str]
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the scan targets of an organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationScanTargets>
-        :param organization_id: the ID of the organization
-        : return: an iterator over the scan target objects
-        """
-        yield from self._request(
-            "GET", f"/organizations/{validate_uuid(organization_id)}/scantargets"
-        ).json()
-
-    def create_organization_scan_target(
-        self,
-        organization_id: Union[UUID, str],
-        kind: ScanTargetKind,
-        name: str,
-        credential: Union[
-            ScanTargetAWS,
-            ScanTargetAZURE,
-            ScanTargetGCP,
-            ScanTargetHUAWEI,
-            ScanTargetDOMAIN,
-            ScanTargetORACLE,
-        ],
-        schedule: Union[str, ScanTargetSchedule] = ScanTargetSchedule.TWENTY_FOUR_HOURS,
-    ) -> Dict:
-        """
-        Create a new scan target in organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/createOrganizationScanTargets>
-        :param organization_id: the ID of the organization
-        :param kind: the Kind of scan target (AWS, GCP, AZURE)
-        :param name: the name of the scan target
-        :param credential: credentials to access the cloud account to be scanned:
-            * For AWS scan targets, provide the account ID in the *account* field
-            * For Azure scan targets, provide *applicationId*, *subscriptionId*, *directoryId* and *secret* fields.
-            * For GCP scan targets, provide a *projectId* field
-            * For DOMAIN scan targets, provide a URL in the *domain* field
-        :param schedule: schedule as a string or enum version of the scan frequency
-        :return: a dict representing the newly created scan target
-        """
-        validate_class(kind, ScanTargetKind)
-        validate_class(name, str)
-
-        if kind == ScanTargetKind.AWS:
-            validate_class(credential, ScanTargetAWS)
-        elif kind == ScanTargetKind.AZURE:
-            validate_class(credential, ScanTargetAZURE)
-        elif kind == ScanTargetKind.GCP:
-            validate_class(credential, ScanTargetGCP)
-        elif kind == ScanTargetKind.HUAWEI:
-            validate_class(credential, ScanTargetHUAWEI)
-        elif kind == ScanTargetKind.DOMAIN:
-            validate_class(credential, ScanTargetDOMAIN)
-        elif kind == ScanTargetKind.ORACLE:
-            validate_class(credential, ScanTargetORACLE)
-
-        body = {
-            "name": name,
-            "kind": kind,
-            "credential": credential,
-            "schedule": ScanTargetSchedule.from_value(schedule).value,
-        }
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets",
-            body=body,
-        ).json()
-
-    def get_organization_scan_target(
-        self, organization_id: Union[UUID, str], scan_target_id: Union[UUID, str]
-    ) -> Dict:
-        """
-        Get scan target of organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationScanTargetById>
-        :param scan_target_id:
-        :param organization_id: the ID of the organization
-        :return: a dict representing the scan target
-        """
-        return self._request(
-            "GET",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets/"
-            f"{validate_uuid(scan_target_id)}",
-        ).json()
-
-    def update_organization_scan_target(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_id: Union[UUID, str],
-        name: str,
-        schedule: Union[str, ScanTargetSchedule],
-    ) -> Dict:
-        """
-        Update scan target of organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/editOrganizationScanTargetById>
-        :param schedule:
-        :param scan_target_id:
-        :param name:
-        :param organization_id: the ID of the organization
-        :return: a dict representing the organization follower
-        """
-
-        body = {
-            "name": name,
-            "schedule": ScanTargetSchedule.from_value(schedule).value,
-        }
-
-        return self._request(
-            "PUT",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets/"
-            f"{validate_uuid(scan_target_id)}",
-            body=body,
-        ).json()
-
-    def delete_organization_scan_target(
-        self, organization_id: Union[UUID, str], scan_target_id: Union[UUID, str]
-    ) -> bool:
-        """
-        Delete scan target of organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/deleteOrganizationScanTargetById>
-        :param organization_id: the ID of the organization
-        :param scan_target_id:
-        :return: a boolean if success
-        """
-        return self._request(
-            "DELETE",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets/"
-            f"{validate_uuid(scan_target_id)}",
-        ).json()
-
-    def start_organization_scan_target_scan(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_id: Union[UUID, str],
-        force: Optional[bool],
-    ) -> bool:
-        """
-        Starts a scan on the specified scan target.
-        <https://api.zanshin.tenchisecurity.com/#operation/scanOrganizationScanTarget>
-        :param organization_id: the ID of organization the scan target belongs to
-        :param scan_target_id: the ID of the scan target
-        :param force: whether to force a scan that is in state NEW or INVALID_CREDENTIAL
-        :return: a boolean if success
-        """
-
-        params = {
-            "force": "true" if force else "false"  # Http params are always strings
-        }
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets/"
-            f"{validate_uuid(scan_target_id)}/scan",
-            params=params,
-        ).json()
-
-    def stop_organization_scan_target_scan(
-        self, organization_id: Union[UUID, str], scan_target_id: Union[UUID, str]
-    ) -> bool:
-        """
-        Stop a scan on the specific scan target
-        :param organization_id: the ID of organization the scan target belongs to
-        :param scan_target_id: the ID of the scan target
-        :return: a boolean if success
-        """
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets/"
-            f"{validate_uuid(scan_target_id)}/stop",
-        ).json()
-
-    def check_organization_scan_target(
-        self, organization_id: Union[UUID, str], scan_target_id: Union[UUID, str]
-    ) -> Dict:
-        """
-        Check scan target.
-        <https://api.zanshin.tenchisecurity.com/#operation/checkOrganizationScanTarget>
-        :param organization_id: the ID of organization the scan target belongs to
-        :param scan_target_id: the ID of the scan target
-        :return: a dict representing the scan target
-        """
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets/"
-            f"{validate_uuid(scan_target_id)}/check",
-        ).json()
-
-    def get_gworkspace_oauth_link(
-        self, organization_id: Union[UUID, str], scan_target_id: Union[UUID, str]
-    ) -> Dict:
-        """
-        Retrieve a link to allow the user to authorize zanshin to read info from their gworkspace environment.
-        <https://api.zanshin.tenchisecurity.com/#operation/getGworkspaceOauthLink>
-        :return: a dict with the link
-        """
-        return self._request(
-            "GET",
-            f"/gworkspace/oauth/link?scanTargetId={validate_uuid(scan_target_id)}"
-            f"&organizationId={validate_uuid(organization_id)}",
-        ).json()
-
-    ###################################################
-    # Organization Scan Target Scan
-    ###################################################
-
-    def iter_organization_scan_target_scans(
-        self, organization_id: Union[UUID, str], scan_target_id: Union[UUID, str]
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the scan of a scan target.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationScanTargetScans>
-        :param organization_id: the ID of the organization
-        :param scan_target_id: the ID of the scan target
-        :return: an iterator over the JSON decoded scans
-        """
-        yield from self._request(
-            "GET",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets/"
-            f"{validate_uuid(scan_target_id)}/scans",
-        ).json().get("data", [])
-
-    def get_organization_scan_target_scan(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_id: Union[UUID, str],
-        scan_id: Union[UUID, str],
-    ) -> Dict:
-        """
-        Get scan of scan target.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationScanTargetScanSlot>
-        :param organization_id: the ID of the organization
-        :param scan_target_id: the ID of the scan target
-        :param scan_id: the ID of the scan
-        :return: a dict representing the scan
-        """
-        return self._request(
-            "GET",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets/"
-            f"{validate_uuid(scan_target_id)}/scans/{scan_id}",
-        ).json()
-
-    ###################################################
-    # Organization Scan Target Groups
-    ###################################################
-
-    def iter_organization_scan_target_groups(
-        self, organization_id: Union[UUID, str]
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the scan targets groups.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationScanTargetGroups>
-        :param organization_id: the ID of the organization
-        : return: an iterator over the scan target groups
-        """
-        yield from self._request(
-            "GET", f"/organizations/{validate_uuid(organization_id)}/scantargetgroups"
-        ).json()
-
-    def get_organization_scan_target_group(
-        self, organization_id: Union[UUID, str], scan_target_group_id: Union[UUID, str]
-    ) -> Dict:
-        """
-        Get scan target group of organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationScanTargetGroupById>
-        :param scan_target_group_id:
-        :param organization_id: the ID of the organization
-        :return: a dict representing the scan target group
-        """
-        return self._request(
-            "GET",
-            f"/organizations/{validate_uuid(organization_id)}/scantargetgroups/"
-            f"{validate_uuid(scan_target_group_id)}",
-        ).json()
-
-    def create_scan_target_group(
-        self, organization_id: Union[UUID, str], kind: ScanTargetKind, name: str
-    ) -> Dict:
-        """
-        Create a new scan target group.
-        <https://api.zanshin.tenchisecurity.com/#operation/createOrganizationScanTargetGroup>
-        :param organization_id: the ID of the organization
-        :param kind: The type of cloud of this scan target group
-        :param name: the name of the scan target group
-        :return: a dict representing the newly created scan target group
-        """
-        validate_class(kind, ScanTargetKind)
-        validate_class(name, str)
-        if kind != ScanTargetKind.ORACLE:
-            raise ValueError(
-                f"{repr(kind.value)} is not accepted. 'ORACLE' is expected"
-            )
-
-        body = {
-            "name": name,
-            "kind": kind,
-        }
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/scantargetgroups",
-            body=body,
-        ).json()
-
-    def update_scan_target_group(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_group_id: Union[UUID, str],
-        name: str,
-    ) -> Dict:
-        """
-        Update scan target group.
-        <https://api.zanshin.tenchisecurity.com/#operation/UpdateOrganizationScanTargetGroup>
-        :param scan_target_group_id: the ID of the scan target group
-        :param name: The scan target group assigned name
-        :param organization_id: the ID of the organization
-        :return: a dict representing the scan target group
-        """
-
-        body = {"name": name}
-
-        return self._request(
-            "PUT",
-            f"/organizations/{validate_uuid(organization_id)}/scantargetgroups/"
-            f"{validate_uuid(scan_target_group_id)}",
-            body=body,
-        ).json()
-
-    def iter_scan_target_group_compartments(
-        self, organization_id: Union[UUID, str], scan_target_group_id: Union[UUID, str]
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the compartments of a scan target group.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationComapartmentsFromScanTargetGroup>
-        :param organization_id: the ID of the organization
-        :param scan_target_group_id: the ID of the scan target group
-        :return: an iterator over the compartments of a scan target group
-        """
-        yield from self._request(
-            "GET",
-            f"/organizations/{validate_uuid(organization_id)}/scantargetgroups/"
-            f"{validate_uuid(scan_target_group_id)}/targets",
-        ).json()
-
-    def get_scan_target_group_script(
-        self, organization_id: Union[UUID, str], scan_target_group_id: Union[UUID, str]
-    ) -> Dict:
-        """
-        Get the terraform download URL of the scan target group.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationScanTargetGroupScrip>
-        :param organization_id: the ID of the organization
-        :param scan_target_group_id: the ID of the scan target group
-        :return: Scan target group terraform URL
-        """
-        return self._request(
-            "GET",
-            f"/organizations/{validate_uuid(organization_id)}/scantargetgroups/"
-            f"{validate_uuid(scan_target_group_id)}/scripts",
-        ).json()
-
-    def iter_scan_targets_from_group(
-        self, organization_id: Union[UUID, str], scan_target_group_id: Union[UUID, str]
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the scan targets of a group.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationScanTargetFromScanTargetGroup>
-        :param organization_id: the ID of the organization
-        :param scan_target_group_id: the ID of the scan target group
-        :return: an iterator over scan targets of a group
-        """
-        yield from self._request(
-            "GET",
-            f"/organizations/{validate_uuid(organization_id)}/scantargetgroups/"
-            f"{validate_uuid(scan_target_group_id)}/scantargets",
-        ).json()
-
-    def delete_organization_scan_target_group(
-        self, organization_id: Union[UUID, str], scan_target_group_id: Union[UUID, str]
-    ) -> bool:
-        """
-        Delete scan target group of organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/deleteOrganizationScanTargetGroupById>
-        :param organization_id: the ID of the organization
-        :param scan_target_group_id:
-        :return: a boolean if success
-        """
-        return self._request(
-            "DELETE",
-            f"/organizations/{validate_uuid(organization_id)}/scantargetgroups/"
-            f"{validate_uuid(scan_target_group_id)}",
-        ).json()
-
-    def insert_scan_target_group_credential(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_group_id: Union[UUID, str],
-        credential: ScanTargetGroupCredentialListORACLE,
-    ) -> Dict:
-        """
-        Insert an already created scan target group.
-        <https://api.zanshin.tenchisecurity.com/#operation/UpdateOrganizationScanTargetGroupCredential>
-        :param organization_id: the ID of the organization
-        :param scan_target_group_id: the ID of the scan target group
-        :param credential: scan target group credential oracle
-        :return: a dict representing scan target group
-        """
-
-        validate_class(credential, ScanTargetGroupCredentialListORACLE)
-
-        body = {
-            "credential": credential,
-        }
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/scantargetgroups/"
-            f"{validate_uuid(scan_target_group_id)}",
-            body=body,
-        ).json()
-
-    def create_scan_target_by_compartments(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_group_id: Union[UUID, str],
-        name: str,
-        ocid: str,
-    ) -> Dict:
-        """
-        Create Scan Targets from previous listed compartments inside the scan target group.
-        <https://api.zanshin.tenchisecurity.com/#operation/createOrganizationScanTargetByCompartments>
-        :param organization_id: the ID of the organization
-        :param scan_target_group_id: the ID of the scan target group
-        :param ocid: Oracle Compartment Id
-        :param name: the name of the scan target group
-        :return: a dict representing the scan target
-        """
-        validate_class(ocid, str)
-        validate_class(name, str)
-
-        compartments = [{"name": name, "ocid": ocid}]
-
-        body = {"compartments": compartments}
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/scantargetgroups/"
-            f"{validate_uuid(scan_target_group_id)}/targets",
-            body=body,
-        ).json()
-
-    ###################################################
-    # Alerts
-    ###################################################
-
-    def _get_alerts_page(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        rule: Optional[str] = None,
-        states: Optional[Iterable[AlertState]] = None,
-        severities: Optional[Iterable[AlertSeverity]] = None,
-        page: int = 1,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        created_at_start: Optional[str] = None,
-        created_at_end: Optional[str] = None,
-        updated_at_start: Optional[str] = None,
-        updated_at_end: Optional[str] = None,
-        search: Optional[str] = None,
-        order: Optional[AlertsOrderOpts] = None,
-        sort: Optional[SortOpts] = None,
-    ) -> Dict:
-        """
-        Internal method to retrieve a single page of alerts from an organization
-        :param organization_id: the ID of the organization
-        :param scan_target_ids: optional list of scan target IDs to list alerts from, defaults to all
-        :param rule: to filter alerts from (rule), not passing the field will fetch all
-        :param states: optional list of states to filter returned alerts, defaults to all
-        :param severities: optional list of severities to filter returned alerts, defaults to all
-        :param page: page number, starting from 1
-        :param page_size: page size of alerts
-        :param language: language the rule will be returned
-        :param created_at_start: Search alerts by creation date - greater or equals than
-        :param created_at_end: Search alerts by creation date - less or equals than
-        :param updated_at_start: Search alerts by update date - greater or equals than
-        :param updated_at_end: Search alerts by update date - less or equals than
-        :param order: Sort order to use (ascending or descending)
-        :param sort: Which field to sort on
-        :return: an iterator over the JSON decoded alerts
-        :return:
-        """
-        validate_int(page, min_value=1, required=True)
-        validate_int(page_size, min_value=1, required=True)
-        body = {
-            "organizationId": validate_uuid(organization_id),
-            "page": page,
-            "pageSize": page_size,
-        }
-        if search:
-            validate_class(search, str)
-            body["search"] = search
-        if order:
-            validate_class(order, AlertsOrderOpts)
-            body["order"] = order.value
-        if sort:
-            validate_class(sort, SortOpts)
-            body["sort"] = sort.value
-        if scan_target_ids:
-            if isinstance(scan_target_ids, str):
-                scan_target_ids = [scan_target_ids]
-            validate_class(scan_target_ids, Iterable)
-            body["scanTargetIds"] = [validate_uuid(x) for x in scan_target_ids]
-        if rule:
-            body["rule"] = rule
-        if states:
-            if isinstance(states, str) or isinstance(states, AlertState):
-                states = [states]
-            validate_class(states, Iterable)
-            body["states"] = [validate_class(x, AlertState).value for x in states]
-        if severities:
-            if isinstance(severities, str):
-                severities = [severities]
-            validate_class(severities, Iterable)
-            body["severities"] = [
-                validate_class(x, AlertSeverity).value for x in severities
-            ]
-        if language:
-            validate_class(language, Languages)
-            body["lang"] = language.value
-        # TODO: Validate these dates.
-        if created_at_start:
-            body["createdAtStart"] = created_at_start
-        if created_at_end:
-            body["createdAtEnd"] = created_at_end
-        if updated_at_start:
-            body["updatedAtStart"] = updated_at_start
-        if updated_at_end:
-            body["updatedAtEnd"] = updated_at_end
-
-        return self._request("POST", "/alerts", body=body).json()
-
-    def iter_alerts(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        rule: Optional[str] = None,
-        states: Optional[Iterable[AlertState]] = None,
-        severities: Optional[Iterable[AlertSeverity]] = None,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        created_at_start: Optional[str] = None,
-        created_at_end: Optional[str] = None,
-        updated_at_start: Optional[str] = None,
-        updated_at_end: Optional[str] = None,
-        search: Optional[str] = None,
-        order: Optional[AlertsOrderOpts] = None,
-        sort: Optional[SortOpts] = None,
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the alerts of an organization by loading them, transparently paginating on the API
-        <https://api.zanshin.tenchisecurity.com/#operation/listAllAlert>
-        :param organization_id: the ID of the organization
-        :param scan_target_ids: optional list of scan target IDs to list alerts from, defaults to all
-        :param rule: to filter alerts from (rule), not passing the field will fetch all
-        :param states: optional list of states to filter returned alerts, defaults to all
-        :param severities: optional list of severities to filter returned alerts, defaults to all
-        :param page_size: the number of alerts to load from the API at a time
-        :param language: language the rule will be returned. Ignored when historical is enabled
-        :param created_at_start: Search alerts by creation date - greater or equals than
-        :param created_at_end: Search alerts by creation date - less or equals than
-        :param updated_at_start: Search alerts by update date - greater or equals than
-        :param updated_at_end: Search alerts by update date - less or equals than
-        :param search: Search string to find in alerts
-        :param order: Sort order to use (ascending or descending)
-        :param sort: Which field to sort on
-        :return: an iterator over the JSON decoded alerts
-        """
-        page = self._get_alerts_page(
-            organization_id,
-            scan_target_ids,
-            rule,
-            states,
-            severities,
-            page=1,
-            page_size=page_size,
-            language=language,
-            created_at_start=created_at_start,
-            created_at_end=created_at_end,
-            updated_at_start=updated_at_start,
-            updated_at_end=updated_at_end,
-            search=search,
-            order=order,
-            sort=sort,
-        )
-        yield from page.get("data", [])
-        for page_number in range(
-            2, int(ceil(page.get("total", 0) / float(page_size))) + 1
-        ):
-            page = self._get_alerts_page(
-                organization_id,
-                scan_target_ids,
-                rule,
-                states,
-                severities,
-                page=page_number,
-                page_size=page_size,
-                language=language,
-                created_at_start=created_at_start,
-                created_at_end=created_at_end,
-                updated_at_start=updated_at_start,
-                updated_at_end=updated_at_end,
-                search=search,
-                order=order,
-                sort=sort,
-            )
-            yield from page.get("data", [])
-
-    def _get_following_alerts_page(
-        self,
-        organization_id: Union[UUID, str],
-        following_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        rule: Optional[str] = None,
-        states: Optional[Iterable[AlertState]] = None,
-        severities: Optional[Iterable[AlertSeverity]] = None,
-        page: int = 1,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        created_at_start: Optional[str] = None,
-        created_at_end: Optional[str] = None,
-        updated_at_start: Optional[str] = None,
-        updated_at_end: Optional[str] = None,
-        search: Optional[str] = None,
-        order: Optional[AlertsOrderOpts] = None,
-        sort: Optional[SortOpts] = None,
-    ) -> Dict:
-        """
-        Internal method to retrieve a single page of alerts from organizations being followed
-        :param organization_id: the ID of the organization
-        :param following_ids: optional list of scan target IDs to list alerts from, defaults to all
-        :param rule: to filter alerts from (rule), not passing the field will fetch all
-        :param states: optional list of states to filter returned alerts, defaults to all
-        :param severities: optional list of severities to filter returned alerts, defaults to all
-        :param page: page number, starting from 1
-        :param page_size: page size of alerts
-        :param language: language the rule will be returned
-        :param created_at_start: Search alerts by creation date - greater or equals than
-        :param created_at_end: Search alerts by creation date - less or equals than
-        :param updated_at_start: Search alerts by update date - greater or equals than
-        :param updated_at_end: Search alerts by update date - less or equals than
-        :param search: Search string to find in alerts
-        :param order: Sort order to use (ascending or descending)
-        :param sort: Which field to sort on
-        :return: the decoded JSON response from the API
-        """
-        validate_int(page, min_value=1, required=True)
-        validate_int(page_size, min_value=1, required=True)
-        body = {
-            "organizationId": validate_uuid(organization_id),
-            "page": page,
-            "pageSize": page_size,
-        }
-
-        if search:
-            validate_class(search, str)
-            body["search"] = search
-        if order:
-            validate_class(order, AlertsOrderOpts)
-            body["order"] = order.value
-        if sort:
-            validate_class(sort, SortOpts)
-            body["sort"] = sort.value
-        if following_ids:
-            if isinstance(following_ids, str):
-                following_ids = [following_ids]
-            validate_class(following_ids, Iterable)
-            body["followingIds"] = [validate_uuid(x) for x in following_ids]
-        if rule:
-            body["rule"] = rule
-        if states:
-            if isinstance(states, str) or isinstance(states, AlertState):
-                states = [states]
-            validate_class(states, Iterable)
-            body["states"] = [validate_class(x, AlertState).value for x in states]
-        if severities:
-            if isinstance(severities, str):
-                severities = [severities]
-            validate_class(severities, Iterable)
-            body["severities"] = [
-                validate_class(x, AlertSeverity).value for x in severities
-            ]
-        if language:
-            validate_class(language, Languages)
-            body["lang"] = language.value
-        if created_at_start:
-            body["CreatedAtStart"] = created_at_start
-        if created_at_end:
-            body["CreatedAtEnd"] = created_at_end
-        if updated_at_start:
-            body["UpdatedAtStart"] = updated_at_start
-        if updated_at_end:
-            body["UpdatedAtEnd"] = updated_at_end
-
-        return self._request("POST", "/alerts/following", body=body).json()
-
-    def iter_following_alerts(
-        self,
-        organization_id: Union[UUID, str],
-        following_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        rule: Optional[str] = None,
-        states: Optional[Iterable[AlertState]] = None,
-        severities: Optional[Iterable[AlertSeverity]] = None,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        created_at_start: Optional[str] = None,
-        created_at_end: Optional[str] = None,
-        updated_at_start: Optional[str] = None,
-        updated_at_end: Optional[str] = None,
-        search: Optional[str] = None,
-        order: Optional[AlertsOrderOpts] = None,
-        sort: Optional[SortOpts] = None,
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the following alerts froms organizations being followed by transparently paginating on the API.
-        <https://api.zanshin.tenchisecurity.com/#operation/listFollowingAlerts>
-        :param organization_id: the ID of the organization
-        :param following_ids: optional list of IDs of organizations you are following to list alerts from, defaults to
-               all
-        :param rule: to filter alerts from (rule), not passing the field will fetch all
-        :param states: optional list of states to filter returned alerts, defaults to all
-        :param severities: optional list of severities to filter returned alerts, defaults to all
-        :param page_size: the number of alerts to load from the API at a time
-        :param language: language the rule will be returned
-        :param created_at_start: Search alerts by creation date - greater or equals than
-        :param created_at_end: Search alerts by creation date - less or equals than
-        :param updated_at_start: Search alerts by update date - greater or equals than
-        :param updated_at_end: Search alerts by update date - less or equals than
-        :param search: Search string to find in alerts
-        :param order: Sort order to use (ascending or descending)
-        :param sort: Which field to sort on
-        :return: an iterator over the JSON decoded alerts
-        """
-
-        page = self._get_following_alerts_page(
-            organization_id,
-            following_ids,
-            rule,
-            states,
-            severities,
-            page=1,
-            page_size=page_size,
-            language=language,
-            created_at_start=created_at_start,
-            created_at_end=created_at_end,
-            updated_at_start=updated_at_start,
-            updated_at_end=updated_at_end,
-            search=search,
-            order=order,
-            sort=sort,
-        )
-        yield from page.get("data", [])
-        for page_number in range(
-            2, int(ceil(page.get("total", 0) / float(page_size))) + 1
-        ):
-            page = self._get_following_alerts_page(
-                organization_id,
-                following_ids,
-                rule,
-                states,
-                severities,
-                page=page_number,
-                page_size=page_size,
-                language=language,
-                created_at_start=created_at_start,
-                created_at_end=created_at_end,
-                updated_at_start=updated_at_start,
-                updated_at_end=updated_at_end,
-                search=search,
-                order=order,
-                sort=sort,
-            )
-            yield from page.get("data", [])
-
-    def _get_alerts_history_page(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        page_size: int = 100,
-        language: Optional[Iterable[Languages]] = None,
-        cursor: Optional[str] = None,
-    ) -> Dict:
-        """
-        Internal method to retrieve a single page of alerts history from an organization
-        :param organization_id: the ID of the organization
-        :param scan_target_ids: optional list of scan target IDs to list alerts from, defaults to all
-        :param page_size: page size of alerts
-        :param language: language the rule will be returned
-        :param cursor: Alert Cursor of the last alert consumed, when this value is passed, subsequent alert histories
-               will be returned.
-        :return: an iterator over the JSON decoded alerts
-        :return:
-        """
-        validate_int(page_size, min_value=1, required=True)
-        body = {"organizationId": validate_uuid(organization_id), "pageSize": page_size}
-
-        if scan_target_ids:
-            if isinstance(scan_target_ids, str):
-                scan_target_ids = [scan_target_ids]
-            validate_class(scan_target_ids, Iterable)
-            body["scanTargetIds"] = [validate_uuid(x) for x in scan_target_ids]
-        if language:
-            validate_class(language, Languages)
-            body["lang"] = language.value
-        if cursor:
-            body["cursor"] = cursor
-
-        return self._request("POST", "/alerts/history", body=body).json()
-
-    def iter_alerts_history(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        cursor: Optional[str] = None,
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the alert's history of an organization by loading them, transparently paginating on the API.
-        <https://api.zanshin.tenchisecurity.com/#operation/listAllAlertsHistory>
-        :param organization_id: the ID of the organization
-        :param scan_target_ids: optional list of scan target IDs to list alerts from, defaults to all
-        :param page_size: the number of alerts to load from the API at a time
-        :param language: language the rule will be returned.
-        :param cursor: Alert Cursor of the last alert consumed, when this value is passed, subsequent alert histories
-               will be returned.
-        :return: an iterator over the JSON decoded alerts
-        """
-
-        page = self._get_alerts_history_page(
-            organization_id,
-            scan_target_ids,
-            page_size=page_size,
-            language=language,
-            cursor=cursor,
-        )
-        data = page.get("data", [])
-        yield from data
-
-        while len(data) > 0:
-            cursor = data[len(data) - 1]["cursor"]
-            page = self._get_alerts_history_page(
-                organization_id,
-                scan_target_ids,
-                page_size=page_size,
-                language=language,
-                cursor=cursor,
-            )
-            data = page.get("data", [])
-            yield from data
-
-    def _get_alerts_following_history_page(
-        self,
-        organization_id: Union[UUID, str],
-        following_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        cursor: Optional[str] = None,
-    ) -> Dict:
-        """
-        Internal method to retrieve a single page of alerts history from an organization
-        :param organization_id: the ID of the organization
-        :param following_ids: optional list of IDs of organizations you are following to list alerts from, defaults to
-               all
-        :param page_size: page size of alerts
-        :param language: language the rule will be returned
-        :param cursor: Alert Cursor of the last alert consumed, when this value is passed, subsequent alert histories
-               will be returned.
-        :return: an iterator over the JSON decoded alerts
-        :return:
-        """
-        validate_int(page_size, min_value=1, required=True)
-        body = {"organizationId": validate_uuid(organization_id), "pageSize": page_size}
-
-        if following_ids:
-            if isinstance(following_ids, str):
-                following_ids = [following_ids]
-            validate_class(following_ids, Iterable)
-            body["followingIds"] = [validate_uuid(x) for x in following_ids]
-        if language:
-            validate_class(language, Languages)
-            body["lang"] = language.value
-        if cursor:
-            body["cursor"] = cursor
-
-        return self._request("POST", "/alerts/history/following", body=body).json()
-
-    def iter_alerts_following_history(
-        self,
-        organization_id: Union[UUID, str],
-        following_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        cursor: Optional[str] = None,
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the alert's history of an organization by loading them, transparently paginating on the API
-        <https://api.zanshin.tenchisecurity.com/#operation/listAllAlertsHistoryFollowing>
-        :param organization_id: the ID of the organization
-        :param following_ids: optional list of IDs of organizations you are following to list alerts from, defaults to
-               all
-        :param page_size: the number of alerts to load from the API at a time
-        :param language: language the rule will be returned. Ignored when historical is enabled
-        :param cursor: Alert Cursor of the last alert consumed, when this value is passed, subsequent alert histories
-               will be returned
-        :return: an iterator over the JSON decoded alerts
-        """
-        page = self._get_alerts_following_history_page(
-            organization_id,
-            following_ids,
-            page_size=page_size,
-            language=language,
-            cursor=cursor,
-        )
-        data = page.get("data", [])
-        yield from data
-
-        while len(data) > 0:
-            cursor = data[len(data) - 1]["cursor"]
-            page = self._get_alerts_following_history_page(
-                organization_id,
-                following_ids,
-                page_size=page_size,
-                language=language,
-                cursor=cursor,
-            )
-            data = page.get("data", [])
-            yield from data
-
-    def _get_grouped_alerts_page(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        states: Optional[Iterable[AlertState]] = None,
-        severities: Optional[Iterable[AlertSeverity]] = None,
-        page: int = 1,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        search: Optional[str] = None,
-        order: Optional[AlertsOrderOpts] = None,
-        sort: Optional[SortOpts] = None,
-    ) -> Dict:
-        """
-        Internal method to retrieve a single page of alerts from an organization
-        :param organization_id: the ID of the organization
-        :param scan_target_ids: optional list of scan target IDs to list alerts from, defaults to all
-        :param states: optional list of states to filter returned alerts, defaults to all
-        :param severities: optional list of severities to filter returned alerts, defaults to all
-        :param page: page number, starting from 1
-        :param page_size: page size of alerts
-        :param language: language to use for the returned rules
-        :param search: Search string to find in alerts
-        :param order: Sort order to use (ascending or descending)
-        :param sort: Which field to sort on
-
-        :return:
-        """
-        validate_int(page, min_value=1, required=True)
-        validate_int(page_size, min_value=1, required=True)
-        body = {
-            "organizationId": validate_uuid(organization_id),
-            "page": page,
-            "pageSize": page_size,
-        }
-        if language:
-            validate_class(language, Languages)
-            body["lang"] = language.value
-        if search:
-            validate_class(search, str)
-            body["search"] = search
-        if order:
-            validate_class(order, AlertsOrderOpts)
-            body["order"] = order.value
-        if sort:
-            validate_class(sort, SortOpts)
-            body["sort"] = sort.value
-        if scan_target_ids:
-            if isinstance(scan_target_ids, str):
-                scan_target_ids = [scan_target_ids]
-            validate_class(scan_target_ids, Iterable)
-            body["scanTargetIds"] = [validate_uuid(x) for x in scan_target_ids]
-        if states:
-            if isinstance(states, str) or isinstance(states, AlertState):
-                states = [states]
-            validate_class(states, Iterable)
-            body["states"] = [validate_class(x, AlertState).value for x in states]
-        if severities:
-            if isinstance(severities, str):
-                severities = [severities]
-            validate_class(severities, Iterable)
-            body["severities"] = [
-                validate_class(x, AlertSeverity).value for x in severities
-            ]
-        return self._request("POST", "/alerts/rules", body=body).json()
-
-    def iter_grouped_alerts(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        states: Optional[Iterable[AlertState]] = None,
-        severities: Optional[Iterable[AlertSeverity]] = None,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        search: Optional[str] = None,
-        order: Optional[AlertsOrderOpts] = None,
-        sort: Optional[SortOpts] = None,
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the grouped alerts of an organization by loading them, transparently paginating on the API.
-        <https://api.zanshin.tenchisecurity.com/#operation/listAllAlertRules>
-        :param organization_id: the ID of the organization
-        :param scan_target_ids: optional list of scan target IDs to list alerts from, defaults to all
-        :param states: optional list of states to filter returned alerts, defaults to all
-        :param severities: optional list of severities to filter returned alerts, defaults to all
-        :param page_size: the number of alerts to load from the API at a time
-        :param language: language to use for the returned rules
-        :param search: Search string to find in alerts
-        :param order: Sort order to use (ascending or descending)
-        :param sort: Which field to sort on
-
-        :return: an iterator over the JSON decoded alerts
-        """
-        page = self._get_grouped_alerts_page(
-            organization_id,
-            scan_target_ids,
-            states,
-            severities,
-            page=1,
-            page_size=page_size,
-            language=language,
-            search=search,
-            order=order,
-            sort=sort,
-        )
-        yield from page.get("data", [])
-        for page_number in range(
-            2, int(ceil(page.get("total", 0) / float(page_size))) + 1
-        ):
-            page = self._get_grouped_alerts_page(
-                organization_id,
-                scan_target_ids,
-                states,
-                severities,
-                page=page_number,
-                page_size=page_size,
-                language=language,
-                search=search,
-                order=order,
-                sort=sort,
-            )
-            yield from page.get("data", [])
-
-    def _get_grouped_following_alerts_page(
-        self,
-        organization_id: Union[UUID, str],
-        following_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        states: Optional[Iterable[AlertState]] = None,
-        severities: Optional[Iterable[AlertSeverity]] = None,
-        page: int = 1,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        search: Optional[str] = None,
-        order: Optional[AlertsOrderOpts] = None,
-        sort: Optional[SortOpts] = None,
-    ) -> Dict:
-        """
-        Internal method to retrieve a single page of alerts from organizations being followed
-        :param organization_id: the ID of the organization
-        :param following_ids: optional list of scan target IDs to list alerts from, defaults to all
-        :param states: optional list of states to filter returned alerts, defaults to all
-        :param severities: optional list of severities to filter returned alerts, defaults to all
-        :param page: page number, starting from 1
-        :param page_size: page size of alerts
-        :param language: language to use for the returned rules
-        :param search: Search string to find in alerts
-        :param order: Sort order to use (ascending or descending)
-        :param sort: Which field to sort on
-        :return: the decoded JSON response from the API
-        """
-        validate_int(page, min_value=1, required=True)
-        validate_int(page_size, min_value=1, required=True)
-        body = {
-            "organizationId": validate_uuid(organization_id),
-            "page": page,
-            "pageSize": page_size,
-        }
-        if language:
-            validate_class(language, Languages)
-            body["lang"] = language.value
-        if search:
-            validate_class(search, str)
-            body["search"] = search
-        if order:
-            validate_class(order, AlertsOrderOpts)
-            body["order"] = order.value
-        if sort:
-            validate_class(sort, SortOpts)
-            body["sort"] = sort.value
-        if following_ids:
-            if isinstance(following_ids, str):
-                following_ids = [following_ids]
-            validate_class(following_ids, Iterable)
-            body["followingIds"] = [validate_uuid(x) for x in following_ids]
-        if states:
-            if isinstance(states, str) or isinstance(states, AlertState):
-                states = [states]
-            validate_class(states, Iterable)
-            body["states"] = [validate_class(x, AlertState).value for x in states]
-        if severities:
-            if isinstance(severities, str):
-                severities = [severities]
-            validate_class(severities, Iterable)
-            body["severities"] = [
-                validate_class(x, AlertSeverity).value for x in severities
-            ]
-
-        return self._request("POST", "/alerts/rules/following", body=body).json()
-
-    def iter_grouped_following_alerts(
-        self,
-        organization_id: Union[UUID, str],
-        following_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        states: Optional[Iterable[AlertState]] = None,
-        severities: Optional[Iterable[AlertSeverity]] = None,
-        page_size: int = 100,
-        language: Optional[Languages] = None,
-        search: Optional[str] = None,
-        order: Optional[AlertsOrderOpts] = None,
-        sort: Optional[SortOpts] = None,
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the grouped following alerts froms organizations being followed by transparently paginating on the
-        API.
-        <https://api.zanshin.tenchisecurity.com/#operation/listAllAlertRulesFollowing>
-        :param organization_id: the ID of the organization
-        :param following_ids: optional list of IDs of organizations you are following to list alerts from, defaults to
-               all
-        :param states: optional list of states to filter returned alerts, defaults to all
-        :param severities: optional list of severities to filter returned alerts, defaults to all
-        :param page_size: the number of alerts to load from the API at a time
-        :param language: language to use for the returned rules
-        :param search: Search string to find in alerts
-        :param order: Sort order to use (ascending or descending)
-        :param sort: Which field to sort on
-        :return: an iterator over the JSON decoded alerts
-        """
-
-        page = self._get_grouped_following_alerts_page(
-            organization_id,
-            following_ids,
-            states,
-            severities,
-            page=1,
-            page_size=page_size,
-            language=language,
-            search=search,
-            order=order,
-            sort=sort,
-        )
-        yield from page.get("data", [])
-        for page_number in range(
-            2, int(ceil(page.get("total", 0) / float(page_size))) + 1
-        ):
-            page = self._get_grouped_following_alerts_page(
-                organization_id,
-                following_ids,
-                states,
-                severities,
-                page=page_number,
-                page_size=page_size,
-                language=language,
-                search=search,
-                order=order,
-                sort=sort,
-            )
-            yield from page.get("data", [])
-
-    def get_alert(self, alert_id: Union[UUID, str]) -> Dict:
-        """
-        Returns the detailed object that describes an alert.
-        <https://api.zanshin.tenchisecurity.com/#operation/getAlertById>
-        :param alert_id: the ID of the alert
-        :return: the decoded JSON object returned by the API
-        """
-        return self._request("GET", f"/alerts/{validate_uuid(alert_id)}").json()
-
-    def iter_alert_history(self, alert_id: Union[UUID, str]) -> Iterator[Dict]:
-        """
-        Iterates over the history of an alert.
-        <https://api.zanshin.tenchisecurity.com/#operation/listAllAlertHistory>
-        :param alert_id: the ID of the alert
-        :return:
-        """
-        yield from self._request(
-            "GET", f"/alerts/{validate_uuid(alert_id)}/history"
-        ).json()
-
-    def iter_alert_comments(self, alert_id: Union[UUID, str]) -> Iterator[Dict]:
-        """
-        Iterates over the comment of an alert.
-        <https://api.zanshin.tenchisecurity.com/#operation/listAllAlertComments>
-        :param alert_id: the ID of the alert
-        :return:
-        """
-        yield from self._request(
-            "GET", f"/alerts/{validate_uuid(alert_id)}/comments"
-        ).json()
-
-    def update_alert(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_id: Union[UUID, str],
-        alert_id: Union[UUID, str],
-        state: Optional[AlertState],
-        labels: Optional[Iterable[str]],
-        comment: Optional[str],
-    ) -> Dict:
-        """
-        Update alert.
-        <https://api.zanshin.tenchisecurity.com/#operation/editOrganizationScanTargetAlertById>
-        :param comment:
-        :param labels:
-        :param state:
-        :param scan_target_id:
-        :param organization_id:
-        :param alert_id: the ID of the alert
-        :return: the decoded JSON object returned by the API
-        """
-
-        body = dict()
-        if state:
-            body["state"] = state
-
-        if labels:
-            body["labels"] = labels
-
-        if comment:
-            body["comment"] = comment
-
-        return self._request(
-            "PUT",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets/"
-            f"{validate_uuid(scan_target_id)}/alerts/{validate_uuid(alert_id)}",
-            body=body,
-        ).json()
-
-    def create_alert_comment(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_id: Union[UUID, str],
-        alert_id: Union[UUID, str],
-        comment: str,
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the comment of an alert.
-        <https://api.zanshin.tenchisecurity.com/#operation/listAllAlertComments>
-        :param comment:
-        :param organization_id: the ID of the organization
-        :param scan_target_id: the ID of the scan target
-        :param alert_id: the ID of the alert
-        :return:
-        """
-
-        body = {"comment": comment}
-
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/scantargets/"
-            f"{validate_uuid(scan_target_id)}/alerts/{validate_uuid(alert_id)}/comments",
-            body=body,
-        ).json()
-
-    ###################################################
-    # Summary
-    ###################################################
-
-    def get_alert_summaries(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        search: Optional[str] = None,
-        language: Optional[Languages] = None,
-    ) -> Dict:
-        """
-        Gets a summary of the current state of alerts for an organization, both in total and broken down by scan
-        target.
-        <https://api.zanshin.tenchisecurity.com/#operation/alertSummary>
-        :param organization_id: the ID of the organization whose alert summaries are desired
-        :param scan_target_ids: optional list of scan target IDs to summarize alerts from, defaults to all
-        :param language: language to use for the returned rules
-        :param search: Search string to find in alerts
-        :return: JSON object containing the alert summaries
-        """
-
-        body = {"organizationId": validate_uuid(organization_id)}
-
-        if scan_target_ids:
-            if isinstance(scan_target_ids, str):
-                scan_target_ids = [scan_target_ids]
-            validate_class(scan_target_ids, Iterable)
-            body["scanTargetIds"] = [validate_uuid(x) for x in scan_target_ids]
-        if search:
-            validate_class(search, str)
-            body["search"] = search
-        if language:
-            validate_class(language, Languages)
-            body["lang"] = language.value
-
-        return self._request("POST", "/alerts/summaries", body=body).json()
-
-    def get_following_alert_summaries(
-        self,
-        organization_id: Union[UUID, str],
-        following_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        search: Optional[str] = None,
-        language: Optional[Languages] = None,
-    ) -> Dict:
-        """
-        Gets a summary of the current state of alerts for followed organizations.
-        <https://api.zanshin.tenchisecurity.com/#operation/alertFollowingSummary>
-        :param organization_id:
-        :param following_ids: list of IDs of organizations being followed to summarize alerts from
-        :param language: language to use for the returned rules
-        :param search: Search string to find in alerts
-        :return: JSON object containing the alert summaries
-        """
-
-        body = {"organizationId": validate_uuid(organization_id)}
-
-        if following_ids:
-            if isinstance(following_ids, str):
-                following_ids = [following_ids]
-            validate_class(following_ids, Iterable)
-            body["followingIds"] = [validate_uuid(x) for x in following_ids]
-        if search:
-            validate_class(search, str)
-            body["search"] = search
-        if language:
-            validate_class(language, Languages)
-            body["lang"] = language.value
-
-        return self._request("POST", "/alerts/summaries/following", body=body).json()
-
-    def get_scan_summaries(
-        self,
-        organization_id: Union[UUID, str],
-        scan_target_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        days: Optional[int] = 7,
-    ) -> Dict:
-        """
-        Returns summaries of scan results over a period of time, summarizing number of alerts that changed states.
-        <https://api.zanshin.tenchisecurity.com/#operation/scanSummary>
-        :param organization_id: the ID of the organization whose scan summaries are desired
-        :param scan_target_ids: optional list of scan target IDs to summarize scans from, defaults to all
-        :param days: number of days to go back in time in historical search
-        :return: JSON object containing the scan summaries
-        """
-
-        body = {
-            "organizationId": validate_uuid(organization_id),
-            "daysBefore": validate_int(days, min_value=1),
-        }
-
-        if scan_target_ids:
-            if isinstance(scan_target_ids, str):
-                scan_target_ids = [scan_target_ids]
-            validate_class(scan_target_ids, Iterable)
-            body["scanTargetIds"] = [validate_uuid(x) for x in scan_target_ids]
-
-        return self._request("POST", "/alerts/summaries/scans", body=body).json()
-
-    def get_following_scan_summaries(
-        self,
-        organization_id: Union[UUID, str],
-        following_ids: Optional[Iterable[Union[UUID, str]]] = None,
-        days: Optional[int] = 7,
-    ) -> Dict:
-        """
-        Gets a summary of the current state of alerts for followed organizations.
-        <https://api.zanshin.tenchisecurity.com/#operation/scanSummaryFollowing>
-        :param organization_id:
-        :param following_ids: list of IDs of organizations being followed to summarize alerts from
-        :param days: number of days to go back in time in historical search
-        :return: JSON object containing the scan summaries
-        """
-
-        body = {
-            "organizationId": validate_uuid(organization_id),
-            "daysBefore": validate_int(days, min_value=1),
-        }
-
-        if following_ids:
-            if isinstance(following_ids, str):
-                following_ids = [following_ids]
-            validate_class(following_ids, Iterable)
-            body["followingIds"] = [validate_uuid(x) for x in following_ids]
-
-        return self._request(
-            "POST", "/alerts/summaries/scans/following", body=body
-        ).json()
-
-    def __repr__(self):
-        return (
-            f"Connection(api_url='{self.api_url}', api_key='***{self._api_key[-6:]}', "
-            f"user_agent='{self.user_agent}', proxy_url='{self._get_sanitized_proxy_url()}')"
-        )
-
-    ###################################################
-    # Onboard Scan Targets
-    ###################################################
-
-    def onboard_scan_target(
-        self,
-        region: str,
-        organization_id: Union[UUID, str],
-        kind: Union[ScanTargetKind, str],
-        name: str,
-        credential: Union[
-            ScanTargetAWS,
-            ScanTargetAZURE,
-            ScanTargetGCP,
-            ScanTargetHUAWEI,
-            ScanTargetDOMAIN,
-        ],
-        boto3_session: any = None,
-        boto3_profile: str = "default",
-        schedule: Union[str, ScanTargetSchedule] = ScanTargetSchedule.TWENTY_FOUR_HOURS,
-    ) -> Dict:
-        """
-        Currently supports only AWS Scan Targets.
-        For AWS Scan Target:
-        If boto3 is installed, creates a Scan Target for the given organization and perform the onboard.
-        :param region: the AWS Region to deploy the CloudFormation Template of Zanshin Service Role.
-        :param organization_id: the ID of the organization to have the new Scan Target.
-        :param kind: the Kind of scan target (AWS, GCP, AZURE, DOMAIN)
-        :param name: the name of the new scan target.
-        :param credential: credentials to access the cloud account to be scanned:
-            * For AWS scan targets, provide the account ID in the *account* field.
-            * For Azure scan targets, provide *applicationId*, *subscriptionId*, *directoryId* and *secret* fields.
-            * For GCP scan targets, provide a *projectId* field.
-            * For DOMAIN scan targets, provide a URL in the *domain* field.
-
-        :param schedule: schedule in string or enum format.
-        :param boto3_profile: boto3 profile name used for CloudFormation Deployment. If none, uses \"default\" profile.
-        :param boto3_session: boto3 session used for CloudFormation Deployment. If informed, will ignore boto3_profile.
-        :return: JSON object containing newly created scan target .
-        """
-
-        if isinstance(kind, str):
-            kind = ScanTargetKind(kind.strip().upper())
-        else:
-            validate_class(kind, ScanTargetKind)
-        self._check_scantarget_is_aws(kind)
-
-        boto3 = self._check_boto3_installation()
-        if not boto3_session:
-            boto3_session = self._get_session_from_boto3_profile(
-                boto3_profile=boto3_profile, boto3=boto3
-            )
-
-        self._check_aws_credentials_are_valid(boto3_session=boto3_session)
-
-        if len(name) < 3:
-            name = f"{name}_{credential['account']}"
-
-        new_scan_target = self.create_organization_scan_target(
-            organization_id, kind, name, credential, schedule
-        )
-        new_scan_target_id = new_scan_target["id"]
-
-        zanshin_stack_name = "tenchi-zanshin-service-role"
-        try:
-            cloudformation_client = self._deploy_cloudformation_zanshin_service_role(
-                boto3_session, region, new_scan_target_id, zanshin_stack_name
-            )
-            retries = 0
-            max_retry = 10
-            wait_between_retries = 10
-            zanshin_stack = self._get_cloudformation_stack_status(
-                zanshin_stack_name, cloudformation_client
-            )
-
-            while zanshin_stack["StackStatus"] != "CREATE_COMPLETE":
-                if not retries:
-                    self._logger.debug(
-                        f"Failed to confirm CloudFormation Stack {zanshin_stack_name} completion. Retrying."
-                    )
-                if retries >= max_retry:
-                    raise RuntimeError("CloudFormation Stack wasn't deployed")
-                time.sleep(wait_between_retries)
-                self._logger.debug(
-                    f"Checking CloudFormation Stack {zanshin_stack_name}..."
-                )
-                retries += 1
-                zanshin_stack = self._get_cloudformation_stack_status(
-                    zanshin_stack_name, cloudformation_client
-                )
-
-        except Exception as error:
-            print("err", error)
-            raise ValueError(
-                f"Failed to confirm CloudFormation Stack {zanshin_stack_name} completion."
-            )
-
-        self.check_organization_scan_target(
-            organization_id=organization_id, scan_target_id=new_scan_target_id
-        )
-        self.start_organization_scan_target_scan(
-            organization_id=organization_id,
-            scan_target_id=new_scan_target_id,
-            force=True,
-        )
-        return self.get_organization_scan_target(
-            organization_id=organization_id, scan_target_id=new_scan_target_id
-        )
-
-    def _deploy_cloudformation_zanshin_service_role(
-        self,
-        boto3_session: object,
-        region: str,
-        new_scan_target_id: str,
-        zanshin_stack_name: str,
-    ):
-        """
-        Instantiate boto3 client for CloudFormation, and create the Stack containing Zanshin Service Role.
-        :return: boto3 cloudformation client.
-        """
-        try:
-            cloudformation_client = boto3_session.client(
-                "cloudformation", region_name=region
-            )
-            cloudformation_client.create_stack(
-                StackName=zanshin_stack_name,
-                TemplateURL="https://s3.amazonaws.com/tenchi-assets/zanshin-service-role.template",
-                Parameters=[
-                    {"ParameterKey": "ExternalId", "ParameterValue": new_scan_target_id}
-                ],
-                Capabilities=["CAPABILITY_IAM", "CAPABILITY_NAMED_IAM"],
-            )
-
-            return cloudformation_client
-        except Exception as e:
-            self._logger.error(
-                "Unable to deploy CloudFormation zanshin-tenchi-service-role. The onboard won't succeed."
-            )
-            raise e
-
-    def _get_cloudformation_stack_status(
-        self, zanshin_stack_name, cloudformation_client
-    ):
-        """
-        Fetch CloudFormation Stack details. Assumes that there's only one Stack with given name.
-        :return: cloudformation stack.
-        """
-        zanshin_stack = cloudformation_client.describe_stacks(
-            StackName=zanshin_stack_name
-        )["Stacks"][0]
-
-        return zanshin_stack
-
-    def _get_session_from_boto3_profile(self, boto3_profile, boto3):
-        """
-        Return boto3_session from boto3_profile informed
-        :return: boto3_session.
-        """
-        return boto3.Session(profile_name=boto3_profile)
-
-    def _check_aws_credentials_are_valid(self, boto3_session):
-        """
-        Check if boto3 informed credentials are valid performing aws sts get-caller-identity. In case of
-        problem, raises ValueError.
-
-        """
-        try:
-            sts = boto3_session.client("sts")
-            sts.get_caller_identity()
-        except Exception as e:
-            self._logger.exception("boto3 session is invalid")
-            raise ValueError(
-                "boto3 session is invalid. Working boto3 session is required."
-            )
-
-    def _check_scantarget_is_aws(self, kind):
-        """
-        Check if informed Scan Target is of AWS Kind. If not, raises NotImplementedError.
-        """
-        if kind != ScanTargetKind.AWS:
-            raise NotImplementedError(
-                f"Onboard doesn't support {kind.value} environment yet"
-            )
-
-    def _check_boto3_installation(self):
-        """
-        Check if boto3 is installed in the current environment. If not, raises ImportError.
-        :return: boto3 module if present.
-        """
-        package_name = "boto3"
-        spec = find_spec(package_name)
-        if spec is None:
-            raise ImportError(
-                f"{package_name} not present. {package_name} is required to perform AWS onboard."
-            )
-
-        module = module_from_spec(spec)
-        sys.modules[package_name] = module
-        spec.loader.exec_module(module)
-        return module
 
 
 def validate_int(
